microservices/volume_of_listened_hours/requirements.txt
py2neo
psycopg2-binary
redis
flask
requests